changed cost column from integer to decimal

+++ migrations/20160623174001_bananas.js
@@ -5,7 +5,7 @@ exports.up = function(knex, Promise) {
        table.integer('userId')
        table.integer('quantity')
        table.string('dateBought')
-       table.integer('cost')
+       table.decimal('cost')
        table.timestamp('timeEntered')
        console.log('bananas table was created!!')
      })
